% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_distance.R
\name{average_distance}
\alias{average_distance}
\alias{average_distance.dts_long}
\title{average_distance}
\usage{
average_distance(x, n)

\method{average_distance}{dts_long}(x, n = 0.5)
}
\arguments{
\item{x}{data read from a dts}

\item{n}{}
}
\description{
average_distance
}
