% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_time.R
\name{average_time}
\alias{average_time}
\alias{average_time.dts_long}
\title{average_time}
\usage{
average_time(x, n)

\method{average_time}{dts_long}(x, n = 30)
}
\arguments{
\item{x}{data read from a dts}

\item{n}{size of the interval to average in seconds}
}
\description{
average_time
}
