% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_variance.R
\name{calculate_variance}
\alias{calculate_variance}
\title{calculate_variance}
\usage{
calculate_variance(dts, col_name = "stokes")
}
\arguments{
\item{col_name}{}
}
\description{
calculate_variance
}
