% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_distance.R
\name{calibration_bath}
\alias{calibration_bath}
\title{#' kmeans_distance
#'
#' @param x 
#' @param n_comp 
#' @param ... 
#'
#' @return
#' @export
#'
#' @examples
#' 
kmeans_distance <- function(x, ...) UseMethod("kmeans_distance")}
\usage{
calibration_bath(x, n_comp = 2)
}
\arguments{
\item{x}{}

\item{n_comp}{}
}
\value{

}
\description{
#' kmeans_distance
#'
#' @param x 
#' @param n_comp 
#' @param ... 
#'
#' @return
#' @export
#'
#' @examples
#' 
kmeans_distance <- function(x, ...) UseMethod("kmeans_distance")
}
\details{
#' @rdname kmeans_distance
#' @export
kmeans_distance.dts_long <- function(x, n_comp = 10, ...) {
  
  
  kmeans_distance(to_matrix(x), n_comp = n_comp)
  
  
}

#' @rdname kmeans_distance
#' @export
kmeans_distance.matrix <- function(x, n_comp = 5, ...) {
  
  
  # find heating times - label ambient, heating, cooling
  m <- to_matrix(x)
  
  bp <- time_breakpoints(x, shift = -20)
  x$trace_time[, type := 'ambient']
  x$trace_time[between(start, bp[1], bp[2]), type := 'heating']
  x$trace_time[start >= bp[2], type := 'cooling']
  
  
  # svd kmeans
  m <- apply(m, 2, scale, scale = FALSE)
  m <- svdr(x, n_comp)$u
  
  # initial clusters
  k <- kmeans(m, centers = 4, iter.max = 30, nstart = 300)
  
  
}
calibration_bath
}
