% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_by_symmetry.R
\name{combine_by_symmetry}
\alias{combine_by_symmetry}
\alias{combine_by_symmetry.numeric}
\alias{combine_by_symmetry.data.table}
\alias{combine_by_symmetry.dts_long}
\title{combine_by_symmetry}
\usage{
combine_by_symmetry(x, ...)

\method{combine_by_symmetry}{numeric}(x)

\method{combine_by_symmetry}{data.table}(x, col_name = "temperature")

\method{combine_by_symmetry}{dts_long}(x, col_name = "temperature")
}
\arguments{
\item{x}{}
}
\description{
combine_by_symmetry
}
