% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_with_temperature.R
\name{compare_with_temperature}
\alias{compare_with_temperature}
\title{compare_with_temperature}
\usage{
compare_with_temperature(x, calib_t, buffer = 0.1, n_rem = 3)
}
\arguments{
\item{x}{}

\item{calib_t}{}

\item{buffer}{temperature difference in degrees}

\item{n_rem}{}

\item{power}{}
}
\description{
compare_with_temperature
}
