% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cor_by_trace}
\alias{cor_by_trace}
\title{correlation for dts matrix by row or by column}
\usage{
cor_by_trace(x, dim)
}
\arguments{
\item{dim}{if 0 then it is by column}
}
\description{
correlation for dts matrix by row or by column
}
\examples{

}
