% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cor_with_trace}
\alias{cor_with_trace}
\title{correlation for dts matrix by row or by column}
\usage{
cor_with_trace(x, y, dim)
}
\arguments{
\item{dim}{}
}
\description{
correlation for dts matrix by row or by column
}
\examples{

}
