% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_with_temperature.R
\name{correlate_with_temperature}
\alias{correlate_with_temperature}
\title{correlate_with_temperature}
\usage{
correlate_with_temperature(x, calib_t, buffer = 0.1, power = 8, n_rem = 3)
}
\arguments{
\item{x}{}

\item{calib_t}{}

\item{buffer}{}

\item{power}{}

\item{n_rem}{}
}
\value{

}
\description{
correlate_with_temperature
}
