% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dts_xml.R
\name{dates_r_to_matlab}
\alias{dates_r_to_matlab}
\title{Convert R datetime to matlab datetime}
\usage{
dates_r_to_matlab(tms)
}
\arguments{
\item{tms}{the POSIXct datetime in R}
}
\value{
datetime in matlab format
}
\description{
Convert R datetime to matlab datetime
}
\examples{
tms <- as.POSIXct('2016-01-01 12:12:10', tz = 'UTC')
dates_r_to_matlab(tms)
 

as.POSIXct((dates_r_to_matlab(tms) - 719529) * 86400, 
origin = "1970-01-01", tz = "UTC")
}
