% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diff_by_trace}
\alias{diff_by_trace}
\title{Difference between dts matrix by row or by column}
\usage{
diff_by_trace(x, shift, dim)
}
\arguments{
\item{x}{matrix}

\item{shift}{how many values to shift}

\item{dim}{by row or by column}
}
\description{
Difference between dts matrix by row or by column
}
