\name{dts-package}
\alias{dts-package}
\alias{dts}
\docType{package}
\title{
\packageTitle{dts}
}
\description{
\packageDescription{dts}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dts}
\packageIndices{dts}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{dts}

Maintainer: \packageMaintainer{dts}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
