% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts-package.R
\docType{package}
\name{dts}
\alias{dts}
\alias{dts-package}
\title{Tools for the analysis of dts}
\description{
This package provides a set of tools for the analysis of dts data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jkennel.github.io/dts}
  \item Report bugs at \url{https://github.com/jkennel/dts/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Kennel \email{jkennel@uoguelph.ca}

Other contributors:
\itemize{
  \item Jonathan Munn \email{jmunn@g360group.org} [research team member]
  \item Carlos Maldaner \email{cmaldaner@g360group.org} [research team member]
  \item Mitchell Brown \email{cmaldaner@g360group.org} [research team member]
  \item Beth Parker \email{bparker@uoguelph.ca} [research team head]
}

}
\keyword{internal}
