% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulated.R
\name{dts_accumulated}
\alias{dts_accumulated}
\title{Accumulated}
\usage{
dts_accumulated(x, dtt = "DateTime", colname = "Value",
  accum = list(Hatch = dttr::dtt_duration(320, "days")))
}
\arguments{
\item{x}{A dts data frame}

\item{dtt}{A string specifying the column with the Date or POSIXct values.}

\item{colname}{A character vector specifying the column(s) with the other values.}

\item{accum}{A uniquely named list do durations.}
}
\value{
The modified DateTime series data frame.
}
\description{
Calculates DateTime when reach accumulated total.
}
\examples{
dts_accumulated(dts_data[1:5,], colname = "Value2", 
  accum = list(Hatch = dttr::dtt_duration(40, "hours")))
}
