% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tibble.R
\name{dts_as_tibble}
\alias{dts_as_tibble}
\title{As Tibble}
\usage{
dts_as_tibble(x, require = FALSE)
}
\arguments{
\item{x}{A data frame to coerce to a tibble.}

\item{require}{A flag indicating whether to require coercion.}
}
\value{
If possible a tibble, otherwise a data frame.
}
\description{
Conditionally converts a data frame to a tibble if the tibble package is installed.
}
\examples{
dts_as_tibble(data.frame())
}
