% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound.R
\name{dts_bound}
\alias{dts_bound}
\title{Bound}
\usage{
dts_bound(x, date_time = "DateTime", value = "Value",
  bounds = c(-Inf, Inf))
}
\arguments{
\item{x}{A dts data frame}

\item{date_time}{A string specifying the column with the Date or POSIXct values.}

\item{value}{A string specifying the column with the double values.}

\item{bounds}{A sorted double vector of the lower and upper permitted bounds.}
}
\value{
A data frame
}
\description{
Values outside the bounds are set to missing values.
}
\examples{
dts_bound(dts_data[1:5,], bounds = c(0,Inf))
}
