% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{dts_complete}
\alias{dts_complete}
\title{Completes a data time series}
\usage{
dts_complete(x, date_time = "DateTime", value = "Value",
  floor = TRUE, sort = TRUE, unique = TRUE,
  units = dttr::dtt_units(x[[date_time]]))
}
\arguments{
\item{x}{A dts data frame}

\item{date_time}{A string specifying the column with the Date or POSIXct values.}

\item{value}{A character vector specifying the column(s) with the other values.}

\item{unique}{A flag indicating whether the values must be unique.}

\item{units}{A string of the units.}
}
\value{
The completed data time series.
}
\description{
Completes a data time series
}
\examples{
dts_complete(dts_data[c(1,3),])
}
