% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{dts_completed}
\alias{dts_completed}
\title{Test if a data time series is complete}
\usage{
dts_completed(x, date_time = "DateTime", value = "Value",
  units = dttr::dtt_units(x[[date_time]]))
}
\arguments{
\item{x}{A dts data frame}

\item{date_time}{A string specifying the column with the Date or POSIXct values.}

\item{value}{A string specifying the column with the double values.}

\item{units}{A string of the time units.}
}
\value{
A flag indicating whether complete.
}
\description{
Test if a data time series is complete
}
\examples{
dts_completed(dts_data[c(1,3),])
}
