% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dts_data}
\alias{dts_data}
\title{dts_data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10325 rows and 4 columns.}
\usage{
dts_data
}
\description{
A data frame of hourly dts data with column DateTime defining the POSIXct values
and column Value specifying the parameter values.
}
\keyword{datasets}
