% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay.R
\name{dts_delay}
\alias{dts_delay}
\title{Delay a data time series}
\usage{
dts_delay(x, dtt = "DateTime", colname = dts_colnames(x), delay = 0,
  units = "days")
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{dtt}{A string specifying the column with the Date or POSIXct values.}

\item{colname}{A character vector specifying the column(s) with the other values.}

\item{delay}{A number (numeric scalar) of the delay.}

\item{units}{A string of the delay units.}
}
\description{
Delay a data time series
}
\examples{
dts_delay(dts_data[1:4,], delay = 2)
}
