% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{dts_density}
\alias{dts_density}
\title{Density}
\usage{
dts_density(x, dtt = "DateTime", colname = "Density", units = "days",
  .fun = "dnorm", .timing, ...)
}
\arguments{
\item{x}{A dts data frame}

\item{dtt}{A string specifying the column with the Date or POSIXct values.}

\item{colname}{A character vector specifying the column(s) with the other values.}

\item{units}{A string of the time units for the variation (3rd and more) 
arguments of .fun.}

\item{.fun}{A string of the name of a function which takes a numeric vector as its first argument 
(the values to calculate the density for), a numeric scalar as it second argument
(the timing of the event) and one or more numeric scalars (representing the
variation in the timing of the event).}

\item{.timing}{A Date or POSIXct vector each element of which 
will be converted into a numeric value to be passed as the second argument to
.fun.}

\item{...}{Additional arguments passed as named arguments to .fun 
which should all be numeric vectors
of the same length as .dtt or numeric scalars (which are reused).}
}
\value{
The modified DateTime series data frame with a column of the density values.
}
\description{
Calculates density by DateTime based on a function.
Its important to realize that the densities are unlikely to sum to zero.
}
\details{
The date time values in x are converted into a numeric vector and 
passed as the first argument to .fun.
}
\examples{
dts_density(dts_data[1:10,], .timing = dts_data$DateTime[2], sd = 2, units = "hours")
}
