% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregate.R
\name{dts_disaggregate}
\alias{dts_disaggregate}
\title{Disaggregate}
\usage{
dts_disaggregate(x, date_time = "DateTime", colname = dts_colnames(x))
}
\arguments{
\item{x}{A dts data frame}

\item{date_time}{A string specifying the column with the Date or POSIXct values.}

\item{colname}{A character vector specifying the column(s) with the other values.}
}
\value{
The disaggregated data frame
}
\description{
Disaggregate
}
\examples{
dts_disaggregate(dts_aggregate(dts_data[1,], units = "years"))
}
