% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate.R
\name{dts_extrapolate}
\alias{dts_extrapolate}
\title{F}
\usage{
dts_extrapolate(x, dtt = "DateTime", colname = dts_colnames(x))
}
\arguments{
\item{x}{A dts data frame}

\item{dtt}{A string specifying the column with the Date or POSIXct values.}

\item{colname}{A character vector specifying the column(s) with the other values.}
}
\value{
A data frame
}
\description{
Fill in missing values by linear interpolation.
}
\examples{
dts_extrapolate(dts_data[2:6,])
}
