% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-dayte.R
\name{dts_fill_dayte}
\alias{dts_fill_dayte}
\title{Fill Date or POSIXct of the Year}
\usage{
dts_fill_dayte(
  x,
  dtt = "DateTime",
  colname = dts_colnames(x),
  min_gap = 0L,
  min_n = 1L,
  feb29_to_28 = TRUE,
  .dts_fun = mean_na_rm,
  ...
)
}
\arguments{
\item{x}{A dts data frame}

\item{dtt}{A string specifying the column with the Date or POSIXct values.}

\item{colname}{A character vector specifying the column(s) with the other
values.}

\item{min_gap}{An integer of the minimum gap to leave between existing values.}

\item{min_n}{A count of the minimum of data points to require.}

\item{feb29_to_28}{A flag indicating whether to treat February 29th as if it is February 28th.}

\item{.dts_fun}{A function which returns a scalar.}

\item{...}{Optional Additional arguments passed to .dts_fun.}
}
\value{
A data frame
}
\description{
Fills in missing values based on Date or DateTime of year.
}
