% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-dayte-time.R
\name{dts_fill_dayte_time}
\alias{dts_fill_dayte_time}
\title{Fill Date or DateTime of the Year}
\usage{
dts_fill_dayte_time(x, date_time = "DateTime", value = "Value",
  min_gap = 0L, min_n = 2L, units = dttr::dtt_units(x[[date_time]]),
  .fun = mean, ...)
}
\arguments{
\item{x}{A dts data frame}

\item{date_time}{A string specifying the column with the Date or POSIXct values.}

\item{value}{A character vector specifying the column(s) with the other values.}

\item{min_gap}{An integer of the minimum gap to leave between existing values.}

\item{min_n}{A count of the minimum of data points to require.}

\item{units}{A string of the units.}

\item{.fun}{A function which returns a scalar.}

\item{...}{Optional Additional arguments passed to .fun.}
}
\value{
A data frame
}
\description{
Fills in missing values based on Date or DateTime of year.
}
\examples{
dts_fill_dayte_time(dts_data, min_gap = 0L)[1:5,]
}
