% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floor.R
\name{dts_floor}
\alias{dts_floor}
\title{Floors the Date or POSIXct column of a date time series data frame.}
\usage{
dts_floor(x, dtt = "DateTime", units = dttr::dtt_units(x[[dtt]]))
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{dtt}{A string specifying the column with the Date or POSIXct values.}

\item{units}{A string of the units.}
}
\value{
The completed data time series.
}
\description{
Floors the Date or POSIXct column of a date time series data frame.
}
\examples{
dts_floor(dts_data[c(1,3),], units = "years")
}
