% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dts_xml.R
\name{dts_format}
\alias{dts_format}
\title{dts_format}
\usage{
dts_format(dts, trim = TRUE)
}
\arguments{
\item{dts}{dts file}

\item{trim}{output to exclude certain distances}
}
\description{
dts_format
}
