% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{dts_interpolate}
\alias{dts_interpolate}
\title{Interpolate}
\usage{
dts_interpolate(x, date_time = "DateTime", colname = dts_colnames(x),
  max_span = .Machine$integer.max)
}
\arguments{
\item{x}{A dts data frame}

\item{date_time}{A string specifying the column with the Date or POSIXct values.}

\item{colname}{A character vector specifying the column(s) with the other values.}

\item{max_span}{An integer of the maximum span to interpolate.
left- and right- continuous step function.}
}
\value{
A data frame
}
\description{
Fill in missing values by linear interpolation.
}
\examples{
dts_interpolate(dts_data[1:5,])
}
