% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dts_interval}
\alias{dts_interval}
\title{Interval}
\usage{
dts_interval(x)
}
\arguments{
\item{x}{A POSIXct object}
}
\value{
A string of the interval
}
\description{
The possible interval values are "seconds", "minutes", "hours",
"days", "months" and "years"
}
\examples{
dts_interval(dts_data$DateTime)
}
