% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.R
\name{dts_lag}
\alias{dts_lag}
\alias{dts_lead}
\alias{dts_delay}
\title{Lag (or lead) columns in a Date Time Series data frame}
\usage{
dts_lag(x, dtt = "DateTime", colname = dts_colnames(x), n = 1L,
  units = dtt_units(x[[dtt]]))

dts_lead(x, dtt = "DateTime", colname = dts_colnames(x), n = 1L,
  units = dtt_units(x[[dtt]]))

dts_delay(x, dtt = "DateTime", colname = dts_colnames(x), n = 1L,
  units = dtt_units(x[[dtt]]))
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{dtt}{A string specifying the column with the Date or POSIXct values.}

\item{colname}{A character vector specifying the column(s) with the other values.}

\item{n}{A int (integer scalar) of the lag.}

\item{units}{A string of the lag time units.}
}
\description{
Lag (or lead) columns in a Date Time Series data frame
}
\examples{
dts_lag(dts_data[1:4,], n = 2)
}
