% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dts_plot}
\alias{dts_plot}
\title{Plot DateTime Series}
\usage{
dts_plot(x, date_time = "DateTime", value = "Value")
}
\arguments{
\item{x}{A data frame with}

\item{date_time}{A string specifying the column with the POSIXct values}

\item{value}{A string specifying the column with the time series values}
}
\value{
A ggplot object
}
\description{
Plot DateTime Series
}
