% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{dts_regress}
\alias{dts_regress}
\title{Regress}
\usage{
dts_regress(x, date_time = "DateTime", value = "Value",
  predictor = "Value2", intercept = TRUE, min_gap = 10L)
}
\arguments{
\item{x}{A dts data frame}

\item{date_time}{A string specifying the column with the Date or POSIXct values.}

\item{value}{A string specifying the column with the double values.}

\item{predictor}{A string of the double column to use as a predictor variable.}

\item{intercept}{A flag indicating whether an intercept should be estimated (or fixed at 0)}

\item{min_gap}{An integer of the minimum gap to leave between regressed and existing values.}
}
\value{
A data frame
}
\description{
Fill in missing values by linear regression.
}
\examples{
dts_regress(dts_data, min_gap = 0L)[1:5,]
}
