% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{dts_separate}
\alias{dts_separate}
\title{Separates the Date or POSIXct column of a data frame into its component parts.}
\usage{
dts_separate(x, date_time = "DateTime", year = TRUE, month = TRUE,
  day = TRUE, hour = FALSE, minute = FALSE, second = FALSE,
  dayte = TRUE, prefix = "", suffix = "")
}
\arguments{
\item{x}{A data frame with a Date or POSIXct column.}

\item{date_time}{A string specifying the column with the Date or POSIXct values.}

\item{year}{A flag indicating whether to add a Year column.}

\item{month}{A flag indicating whether to add a Month column.}

\item{day}{A flag indicating whether to add a Day column.}

\item{hour}{A flag indicating whether to add a Hour column.}

\item{minute}{A flag indicating whether to add a Minute column.}

\item{second}{A flag indicating whether to add a Second column.}

\item{dayte}{A flag indicating whether to add a Dayte column.}

\item{prefix}{A string to add to the start of the column names.}

\item{suffix}{A string to add to the end of the column names.}
}
\value{
The updated data frame.
}
\description{
Separates the Date or POSIXct column of a data frame into its component parts.
}
\examples{
dts_separate(dts_data[c(1,3),])
}
