% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{dts_separate_date}
\alias{dts_separate_date}
\title{Completes the date_time column of a data frame into its component parts.}
\usage{
dts_separate_date(x, date_time = "DateTime", year = TRUE,
  month = TRUE, day = TRUE, dayte = FALSE, doy = FALSE,
  prefix = "", suffix = "")
}
\arguments{
\item{x}{A data frame with a Date or POSIXct column.}

\item{date_time}{A string specifying the column with the Date or POSIXct values.}

\item{year}{A flag indicating whether to add a Year column.}

\item{month}{A flag indicating whether to add a Month column.}

\item{day}{A flag indicating whether to add a Day column.}

\item{dayte}{A flag indicating whether to add a Dayte column.}

\item{doy}{A flag indicating whether to add a Doy column.}

\item{prefix}{A string to add to the start of the column names.}

\item{suffix}{A string to add to the end of the column names.}
}
\value{
The updated data frame.
}
\description{
Completes the date_time column of a data frame into its component parts.
}
\examples{
dts_separate_date(dts_data[c(1,3),])
}
