% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dts_xml.R
\name{dts_to_long}
\alias{dts_to_long}
\title{dts_to_long}
\usage{
dts_to_long(dts, trim_distances = TRUE)
}
\arguments{
\item{dts}{list read from read_dts_xml}

\item{trim_distances}{trim distances}
}
\value{
list of data.tables
}
\description{
dts_to_long
}
