% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dts_xml.R
\name{dts_to_matlab}
\alias{dts_to_matlab}
\title{read_dts_xml}
\usage{
dts_to_matlab(in_dir, out_name, n_cores = 4)
}
\arguments{
\item{in_dir}{path to the input folder}

\item{out_name}{path to the output file}

\item{n_cores}{number of cores to use for parallel processing}
}
\value{
dts list of results
}
\description{
read_dts_xml
}
