% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dts_year}
\alias{dts_year}
\alias{dts_month}
\alias{dts_day}
\alias{dts_hour}
\alias{dts_minute}
\alias{dts_second}
\title{Year, Month, Day, Hour, Minute or Second}
\usage{
dts_year(x)

dts_month(x)

dts_day(x)

dts_hour(x)

dts_minute(x)

dts_second(x)
}
\arguments{
\item{x}{A POSIXct object}
}
\value{
An integer vector
}
\description{
Year, Month, Day, Hour, Minute or Second
}
\section{Functions}{
\itemize{
\item \code{dts_month}: Month

\item \code{dts_day}: Day

\item \code{dts_hour}: Hour

\item \code{dts_minute}: Minute

\item \code{dts_second}: Second
}}

\examples{
dts_year(dts_data$DateTime[1:10])
dts_month(dts_data$DateTime[1:10])
dts_day(dts_data$DateTime[1:10])
dts_hour(dts_data$DateTime[1:10])
dts_minute(dts_data$DateTime[1:10])
dts_second(dts_data$DateTime[1:10])
}
