% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_noise.R
\name{filter_noise}
\alias{filter_noise}
\alias{filter_noise.numeric}
\alias{filter_noise.data.table}
\title{filter_noise}
\usage{
filter_noise(x, n)

\method{filter_noise}{numeric}(x, n)

\method{filter_noise}{data.table}(x, n)
}
\arguments{
\item{x}{data read from a dts}

\item{n}{number of samples}
}
\value{

}
\description{
filter_noise
}
