% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_convolve.R
\name{fit_convolve}
\alias{fit_convolve}
\title{fit_convolve}
\usage{
fit_convolve(x, cool_mult = 0.5, time_var = "start", n_knots = NULL)
}
\arguments{
\item{x}{the dts object}

\item{cool_mult}{currently not used}

\item{time_var}{column name of the that contains the date time info}

\item{n_knots}{number late time knots in lag model - generally keep small}
}
\value{
table of slopes and responses
}
\description{
fit_convolve
}
