% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_convolve.R
\name{fit_convolve2}
\alias{fit_convolve2}
\title{fit_convolve2}
\usage{
fit_convolve2(x, time_var = "start", n_knots = 15)
}
\arguments{
\item{x}{the dts object}

\item{time_var}{column name of the that contains the date time info}

\item{n_knots}{number late time knots in lag model - generally keep small}

\item{cool_mult}{currently not used}
}
\value{
table of slopes and responses
}
\description{
fit_convolve2
}
