% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_traces.R
\name{fit_heating_cooling}
\alias{fit_heating_cooling}
\title{fit_heating_cooling}
\usage{
fit_heating_cooling(
  x,
  start_time = 0.5,
  end_time = 86400,
  heating_type = "heating",
  include_intercept = FALSE
)
}
\arguments{
\item{x}{}

\item{heating_type}{}

\item{include_intercept}{}

\item{start}{}

\item{end}{}
}
\description{
fit_heating_cooling
}
