% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R, R/write_dts.R
\name{generate_report}
\alias{generate_report}
\alias{generate_report.character}
\alias{generate_report.dts_long}
\alias{write_dts.dts_long}
\title{generate_report}
\usage{
generate_report(files_dir, output_dir, ...)

\method{generate_report}{character}(files_dir, output_dir)

\method{generate_report}{dts_long}(files_dir, output_dir)

\method{write_dts}{dts_long}(x, output_dir = getwd(), ...)
}
\arguments{
\item{files_dir}{}

\item{output_dir}{}
}
\value{

}
\description{
generate_report
}
