% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distances.R
\name{get_distances}
\alias{get_distances}
\title{get_distances}
\usage{
get_distances(xml_struct, type)
}
\arguments{
\item{file_path}{path to the file}
}
\value{
distances
}
\description{
get_distances
}
