% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dts_type.R
\name{get_log_name_from_type}
\alias{get_log_name_from_type}
\title{get_log_name_from_type}
\usage{
get_log_name_from_type(type)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
get_log_name_from_type
}
