% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_type.R
\name{get_time_type}
\alias{get_time_type}
\alias{get_time_type.data.table}
\alias{get_time_type.dts_long}
\title{get_time_type}
\usage{
get_time_type(
  x,
  time_type = "heating",
  start_time = NULL,
  end_time = NULL,
  ...
)

\method{get_time_type}{data.table}(
  x,
  time_type = "heating",
  start_time = NULL,
  end_time = NULL,
  ...
)

\method{get_time_type}{dts_long}(
  x,
  time_type = "heating",
  start_time = NULL,
  end_time = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
get_time_type
}
