% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_times.R
\name{get_times}
\alias{get_times}
\alias{get_times.dts}
\alias{get_times.dts_long}
\title{get_times}
\usage{
get_times(x, type = "start", ...)

\method{get_times}{dts}(x, type = "start")

\method{get_times}{dts_long}(x, type = "start")
}
\arguments{
\item{x}{}

\item{type}{'start', 'end', or 'mid'}
}
\value{

}
\description{
get_times
}
\examples{

}
