% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_helpers.R
\name{get_value_xml}
\alias{get_value_xml}
\title{get_value_xml}
\usage{
get_value_xml(search_string, pattern = "(?<=>).*?(?=<)")
}
\arguments{
\item{search_string}{}

\item{pattern}{}
}
\description{
get_value_xml
}
