% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_time.R
\name{group_time}
\alias{group_time}
\title{group_time}
\usage{
group_time(x, interval)
}
\arguments{
\item{x}{time vector that can be coerced to numeric}

\item{interval}{size of interval in time units}
}
\value{
a vector where each time value is assigned the interval midpoint
}
\description{
group_time
}
