% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_by_symmetry.R
\name{half_data}
\alias{half_data}
\alias{half_data.dts_long}
\title{half_data}
\usage{
half_data(x, type = "head", ...)

\method{half_data}{dts_long}(x, type = "head")
}
\arguments{
\item{x}{type read from a dts}
}
\description{
half_data
}
