% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_distance.R
\name{heated_groups}
\alias{heated_groups}
\title{heated_groups}
\usage{
heated_groups(x, n_comp = 5, n_clust = 5, n_buffer = 1000)
}
\arguments{
\item{x}{}

\item{n_comp}{}

\item{n_buffer}{}
}
\description{
heated_groups
}
