% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heating_distance.R
\name{heating_distance}
\alias{heating_distance}
\title{heating_distance}
\usage{
heating_distance(
  x,
  n = 10,
  temperature_change = 0.5,
  power = 5,
  heating_type = "both"
)
}
\arguments{
\item{x}{}

\item{n}{number of points after heating starts}

\item{temperature_change}{temperature change required to include.}

\item{heating_type}{}
}
\description{
heating_distance
}
