% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heating_time.R
\name{heating_time}
\alias{heating_time}
\title{heating_time}
\usage{
heating_time(x, n = -1, heating_type = "both", distance_range = NULL)
}
\arguments{
\item{x}{the DTS dataset}

\item{n}{the spacing between samples in time to calculate the difference.
Should be a negative number.}

\item{heating_type}{the type of test.  Can be "heating" or "both".}
}
\value{

}
\description{
heating_time
}
