% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_to_ultima.R
\name{interpolate_to_ultima}
\alias{interpolate_to_ultima}
\title{interpolate_to_ultima}
\usage{
interpolate_to_ultima(x, ultima_input)
}
\arguments{
\item{x}{the XT DTS data as a list of data tables}

\item{ultima_input}{is the directory of ultima data to interpolate to. Must be .RDS file}
}
\value{

}
\description{
interpolate_to_ultima
}
