% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_time.R
\name{log_average_time}
\alias{log_average_time}
\alias{log_average_time.dts_long}
\title{log_average_time}
\usage{
log_average_time(x, time_column = "log_elapsed_time", n = 100)

\method{log_average_time}{dts_long}(x, time_column = "log_elapsed_time", n = 100)
}
\arguments{
\item{x}{data read from a dts}

\item{n}{}
}
\description{
log_average_time
}
