% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_time.R
\name{log_group_time}
\alias{log_group_time}
\title{log_group_time}
\usage{
log_group_time(x, interval)
}
\arguments{
\item{x}{time vector that can be coerced to numeric}

\item{interval}{size of interval in time units}
}
\value{
a vector where each time value is assigned the interval midpoint
}
\description{
log_group_time
}
