% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_by_symmetry.R
\name{match_region}
\alias{match_region}
\title{match_region}
\usage{
match_region(x, probs = c(0.05, 0.1), resolution_sub = 0.01)
}
\arguments{
\item{x}{}

\item{probs}{}

\item{resolution_sub}{}
}
\value{

}
\description{
match_region
}
