% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_na_rm}
\alias{mean_na_rm}
\title{Arithmetic Mean Without Missing Values}
\usage{
mean_na_rm(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
The arithmetic mean of x after removing missing values.
}
\description{
Arithmetic Mean Without Missing Values
}
\examples{
mean_na_rm(c(1:2, NA))
}
