% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_distances}
\alias{plot_distances}
\title{plot_distances}
\usage{
plot_distances(dts, n = 9, trim_max = 120, trim_min = -5, log_x = FALSE)
}
\arguments{
\item{dts}{the dts set}

\item{n}{number of traces to plot}

\item{trim_max}{maximum temperature values to set to NA}

\item{trim_min}{minimum temperature values to set to NA}
}
\value{
a plotly heatmap
}
\description{
plot_distances
}
