% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(dts, trim_max = 120, trim_min = -5, reverse_axis = TRUE)
}
\arguments{
\item{dts}{the dts set}

\item{trim_max}{maximum temperature values to set to NA}

\item{trim_min}{minimum temperature values to set to NA}
}
\value{
a plotly heatmap
}
\description{
plot_heatmap
}
