% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pnorm_arrival_departure}
\alias{pnorm_arrival_departure}
\title{Normal Arrival-Departure Distribution Function}
\usage{
pnorm_arrival_departure(x, mean = 0, sd = 1, residence = 6)
}
\arguments{
\item{x}{A numeric vector of the times to calculate the distribution for.}

\item{mean}{A number of the mean arrival time.}

\item{sd}{A non-negative number of the standard deviation of the arrival timing.}

\item{residence}{A number of the residence time.}
}
\value{
A numeric vector of the proportion present at x.
}
\description{
Normal Arrival-Departure Distribution Function
}
\examples{
x <- seq(-5,7,by = 0.1) 
plot(x, pnorm_arrival_departure(x), type = "l")
}
