% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_distance.R
\name{process_classification}
\alias{process_classification}
\title{process_classification}
\usage{
process_classification(x, k = 31)
}
\arguments{
\item{x}{vector of classification values}

\item{k}{length of median filter}
}
\value{

}
\description{
process_classification
}
