% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_heating.R
\name{process_heating}
\alias{process_heating}
\title{process_heating}
\usage{
process_heating(dts, set_back = 0, heating = "heating", subset_by = "heated")
}
\arguments{
\item{x}{the DTS dataset}

\item{n}{the spacing between samples in time to calculate the difference.
Should be a negative number.}

\item{heating_type}{the type of test.  Can be "heating" or "both".}
}
\value{

}
\description{
This function takes a DTS dataset and performs rudimentary processing steps to calibrate and find heating times and distances.
}
