% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_distance.R
\name{process_terminal}
\alias{process_terminal}
\title{process_terminal}
\usage{
process_terminal(x, k = 15, fill_value = NA_real_)
}
\arguments{
\item{x}{vector of classification values}

\item{k}{length of median filter}

\item{fill_value}{what to fill the values with}
}
\value{

}
\description{
process_terminal
}
