% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dts_xml.R
\name{read_dts}
\alias{read_dts}
\title{read_dts}
\usage{
read_dts(in_path, n_cores = 1)
}
\arguments{
\item{in_path}{location of the xml files}

\item{n_cores}{specify the number of cores}
}
\description{
read_dts
}
