% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dts_xml.R
\name{read_dts_folder}
\alias{read_dts_folder}
\title{read_dts_folder}
\usage{
read_dts_folder(in_dir, n_cores)
}
\arguments{
\item{in_dir}{location of the xml files}

\item{n_cores}{specify the number of cores}

\item{out_dir}{location of the xml files}
}
\description{
read_dts_folder
}
