% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dts_xml_2.R
\name{read_dts_xml_2}
\alias{read_dts_xml_2}
\title{read_dts_xml_2}
\usage{
read_dts_xml_2(
  in_dir,
  out_dir = getwd(),
  n_cores = 1,
  max_files = Inf,
  return_stokes = TRUE,
  in_memory = TRUE
)
}
\arguments{
\item{in_dir}{path to the input folder}

\item{n_cores}{number of cores to use for parallel processing}

\item{max_files}{maximum number of files to read}

\item{return_stokes}{return stokes and antistokes}

\item{in_memory}{write to disk or return list of data}

\item{date_format}{'matlab' or 'R' format}
}
\value{
dts list of results
}
\description{
read_dts_xml_2
}
