% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dts_xml_3.R
\name{read_dts_xml_3}
\alias{read_dts_xml_3}
\title{read_dts_xml_3}
\usage{
read_dts_xml_3(
  in_dir,
  out_dir = getwd(),
  n_cores = 1,
  max_files = Inf,
  return_stokes = FALSE,
  in_memory = FALSE,
  output_rds = FALSE,
  trim = TRUE,
  time_aggregate_interval = 60L
)
}
\arguments{
\item{time_aggregate_interval}{}
}
\description{
read_dts_xml_3
}
