% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dts_xml_3.R
\name{read_dts_xml_4}
\alias{read_dts_xml_4}
\title{read_dts_xml_4}
\usage{
read_dts_xml_4(
  file_path,
  n_cores = 1L,
  max_files = Inf,
  return_stokes = FALSE,
  trim = TRUE
)
}
\arguments{
\item{trim}{}
}
\description{
read_dts_xml_4
}
