% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dts_xml.R
\name{read_dts_zip}
\alias{read_dts_zip}
\title{read_dts_zip}
\usage{
read_dts_zip(file_path, ...)
}
\arguments{
\item{file}{path to the zip}

\item{n_cores}{number of cores to use for parallel processing}

\item{date_format}{'matlab' or 'R' format}
}
\value{
dts list of results
}
\description{
read_dts_zip
}
