% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_distance.R
\name{resample_distance}
\alias{resample_distance}
\title{resample_distance}
\usage{
resample_distance(x, y, ...)
}
\arguments{
\item{x}{a dts object that you want to spatially resample}

\item{y}{the dts object or vector of distances for the resampling}

\item{...}{arguments to pass to approx}
}
\value{
a dts object that has been resampled
}
\description{
resample_distance
}
