% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{residual_variance_with_trace}
\alias{residual_variance_with_trace}
\title{prediction for dts matrix by row or by column}
\usage{
residual_variance_with_trace(x, y, dim)
}
\arguments{
\item{dim}{}
}
\description{
prediction for dts matrix by row or by column
}
