% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_distance.R
\name{sample_distance}
\alias{sample_distance}
\alias{sample_distance.dts_long}
\title{sample_distance}
\usage{
sample_distance(x, n_traces = 100, ...)

\method{sample_distance}{dts_long}(x, n_traces = 100)
}
\arguments{
\item{x}{data read from a dts}

\item{...}{}
}
\description{
sample_distance
}
