% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_distance.R
\name{sample_times}
\alias{sample_times}
\alias{sample_times.dts_long}
\title{sample_times}
\usage{
sample_times(x, n_traces = 100, ...)

\method{sample_times}{dts_long}(x, n_traces = 100)
}
\arguments{
\item{x}{data read from a sample_times}

\item{...}{}
}
\value{

}
\description{
sample_times
}
