% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_distance}
\alias{select_distance}
\alias{select_distance.dts_long}
\alias{select_distance.data.table}
\title{select_distance}
\usage{
select_distance(x, distances = NULL, by = NULL, ...)

\method{select_distance}{dts_long}(x, distances = NULL, by = NULL)

\method{select_distance}{data.table}(x, distances = NULL)
}
\arguments{
\item{x}{data read from a dts}

\item{...}{}
}
\description{
select_distance
}
