% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_time}
\alias{select_time}
\alias{select_time.dts_long}
\alias{select_time.data.table}
\title{select_time}
\usage{
select_time(x, times, ...)

\method{select_time}{dts_long}(x, times)

\method{select_time}{data.table}(x, times)
}
\arguments{
\item{x}{data read from a dts}

\item{...}{}
}
\description{
select_time
}
