% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_time_ind}
\alias{select_time_ind}
\alias{select_time_ind.dts_long}
\title{select_time_ind}
\usage{
select_time_ind(x, inds, ...)

\method{select_time_ind}{dts_long}(x, inds)
}
\arguments{
\item{x}{data read from a dts}

\item{...}{}
}
\description{
select_time_ind
}
