% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_values_for_distance.R
\name{set_values_for_distance}
\alias{set_values_for_distance}
\title{set_values_for_distance}
\usage{
set_values_for_distance(
  x,
  variable_name,
  begin,
  end,
  specific_depths = NA_real_,
  set_to = TRUE
)
}
\arguments{
\item{x}{}

\item{variable_name}{}

\item{specific_depths}{}

\item{set_to}{}

\item{start_distance}{}

\item{end_distance}{}
}
\value{

}
\description{
set_values_for_distance
}
