% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_values_for_time.R
\name{set_values_for_time}
\alias{set_values_for_time}
\title{set_values_for_distance}
\usage{
set_values_for_time(
  x,
  variable_name,
  begin,
  end,
  specific_times = NA_real_,
  set_to = TRUE
)
}
\arguments{
\item{set_to}{}
}
\description{
set_values_for_distance
}
