% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_values_in_range.R
\name{set_values_in_range}
\alias{set_values_in_range}
\title{set_values_in_range}
\usage{
set_values_in_range(
  x,
  variable_name,
  start_distance,
  end_distance,
  set_to = TRUE
)
}
\arguments{
\item{x}{}

\item{variable_name}{}

\item{start_distance}{}

\item{end_distance}{}

\item{set_to}{}
}
\value{

}
\description{
set_values_in_range
}
