% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_distance_type.R
\name{split_distance_type}
\alias{split_distance_type}
\alias{split_distance_type.dts_long}
\title{split_distance_type}
\usage{
split_distance_type(x, type_col = NULL, ...)

\method{split_distance_type}{dts_long}(x, type_col = NULL)
}
\arguments{
\item{x}{data read from a dts}

\item{...}{}
}
\description{
split_distance_type
}
