% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_up_down.R
\name{split_up_down}
\alias{split_up_down}
\alias{split_up_down.data.table}
\alias{split_up_down.dts_long}
\title{split_up_down
sets the zero point for a well and recalculates distance based on the start,
termination and end of the well's cable}
\usage{
split_up_down(x, begin, termination, end, ...)

\method{split_up_down}{data.table}(x, begin, termination, end, ...)

\method{split_up_down}{dts_long}(x, begin, termination, end, ...)
}
\description{
split_up_down
sets the zero point for a well and recalculates distance based on the start,
termination and end of the well's cable
}
