% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_distance}
\alias{subset_distance}
\alias{subset_distance.dts_long}
\alias{subset_distance.data.table}
\title{subset_distance}
\usage{
subset_distance(x, begin, end, ...)

\method{subset_distance}{dts_long}(x, begin, end)

\method{subset_distance}{data.table}(x, begin, end)
}
\arguments{
\item{x}{data read from a dts}

\item{...}{}
}
\value{

}
\description{
subset_distance
}
