% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R, R/subset.R
\name{select_time}
\alias{select_time}
\alias{subset_time}
\alias{subset_time.dts_long}
\alias{subset_time.data.table}
\title{select_time}
\usage{
select_time(x, times, ...)

subset_time(x, begin, end, ...)

\method{subset_time}{dts_long}(x, begin, end)

\method{subset_time}{data.table}(x, begin, end)
}
\arguments{
\item{x}{data read from a dts}

\item{...}{}
}
\description{
select_time

subset_time
}
