% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_heating_cooling.R
\name{summarize_heating_cooling}
\alias{summarize_heating_cooling}
\title{summarize_heating_cooling}
\usage{
summarize_heating_cooling(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
summarize_heating_cooling
}
