% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermal_conductivity.R
\name{thermal_conductivity}
\alias{thermal_conductivity}
\title{thermal_conductivity}
\usage{
thermal_conductivity(x, power)
}
\arguments{
\item{x}{the slope of the temperature curve}

\item{power}{the input power to the heating cable}
}
\value{

}
\description{
thermal_conductivity
}
