% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_breakpoints.R
\name{time_breakpoints}
\alias{time_breakpoints}
\alias{time_breakpoints.default}
\alias{time_breakpoints.matrix}
\title{time_breakpoints}
\usage{
time_breakpoints(x, shift, col_name, type)

\method{time_breakpoints}{default}(x, shift = -1, col_name = "temperature", type = "both")

\method{time_breakpoints}{matrix}(x, shift = -1, col_name = "temperature", type = "both")
}
\arguments{
\item{x}{data read from a dts}

\item{shift}{distance between differences in number of samples}

\item{col_name}{}

\item{log_type}{one of heating, cooling, or both}
}
\description{
time_breakpoints
}
