% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_data_table.R
\name{to_data_table}
\alias{to_data_table}
\title{to_data_table}
\usage{
to_data_table(x, xt = FALSE, ultima_input = NULL)
}
\arguments{
\item{x}{the DTS data set}

\item{xt}{is set to FALSE by default. If TRUE, the data is to be interpolated and ultima_input must be specified.}

\item{ultima_input}{is only specified when xt is TRUE for interpolation.}
}
\value{

}
\description{
This function takes a DTS data set after heating times and distances are isolated and adds a column of elapsed_time and log_elapsed time. If the dataset is xt and is to be interpolated, argument 'xt' is to be set TRUE and an input for 'ultima_inpue' is specified.
}
