% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_duckdb.R
\name{to_duckdb}
\alias{to_duckdb}
\title{to_duckdb}
\usage{
to_duckdb(x, dbdir, ...)
}
\arguments{
\item{x}{data read from a dts}

\item{dbdir}{name of output file location}

\item{...}{arguments to pass to read_dts_xml_3 time_aggregate_interval,
n_cores, trim, return_stokes, max_files, in_memory}
}
\value{
saves a duckdb database
}
\description{
to_duckdb
}
