% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dts.R
\name{write_dts_to_wellcad}
\alias{write_dts_to_wellcad}
\alias{write_dts_to_wellcad.dts_long}
\title{write_dts_to_wellcad}
\usage{
write_dts_to_wellcad(x, output_dir, ...)

\method{write_dts_to_wellcad}{dts_long}(x, output_dir = getwd(), max_columns = 1000, id = "", ...)
}
\arguments{
\item{...}{}

\item{max_columns}{maximum columns that wellcad can handle}
}
\description{
write_dts_to_wellcad
}
