% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dts_xml.R
\name{write_dts_xml}
\alias{write_dts_xml}
\title{write_dts_xml}
\usage{
write_dts_xml(in_dir, out_dir, n_cores)
}
\arguments{
\item{in_dir}{input directory}

\item{out_dir}{output directory}

\item{n_cores}{how many cores}
}
\description{
write_dts_xml
}
