% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{bath_calibration}
\alias{bath_calibration}
\alias{bath_calibration.dts_long}
\title{bath_calibration}
\usage{
bath_calibration(x, smooth = TRUE, ...)

\method{bath_calibration}{dts_long}(x, smooth = TRUE, ...)
}
\arguments{
\item{x}{data read from a dts}
}
\value{

}
\description{
bath_calibration
}
