% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dts}
\alias{check_dts}
\title{Check DateTime-Series Data Frame}
\usage{
check_dts(x, date_time = "DateTime", value = "Value", nrow = NA,
  key = "DateTime", sorted = FALSE, complete = TRUE,
  x_name = substitute(x), error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{date_time}{A string specifying the column with the POSIXct values.}

\item{value}{A string specifying the column with the numeric values.}

\item{nrow}{A flag indicating whether x should have rows (versus no rows) or a missing value indicating no requirements or a count or count range of the number of rows.}

\item{key}{A character vector of the columns that represent a unique key.}

\item{sorted}{A flag indicating whether the DateTime values must be sorted (in ascending order).}

\item{complete}{A flag indicating whether the DateTime values must be complete.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
A dts data frame is a data frame with a column of non-missing POSIXct values
and a column of numeric values.
}
\examples{
check_dts(dts_data)
}
