% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dts}
\alias{check_dts}
\title{Check DateTime-Series Data Frame}
\usage{
check_dts(
  x,
  dtt = "DateTime",
  colname = character(0),
  nrow = NA,
  nas = TRUE,
  floored = TRUE,
  sorted = FALSE,
  unique = FALSE,
  complete = FALSE,
  units = dttr::dtt_units(x[[dtt]]),
  tz = dttr::dtt_tz(x[[dtt]]),
  exclusive = FALSE,
  order = FALSE,
  x_name = NULL,
  error = TRUE
)
}
\arguments{
\item{x}{A dts data frame}

\item{dtt}{A string specifying the column with the Date or POSIXct values.}

\item{colname}{A character vector specifying the column(s) with the other values.}

\item{nrow}{A flag indicating whether x should have rows (versus no rows) or a missing value indicating no requirements or a count or count range of the number of rows.

@return An invisible copy of x (if it doesn't throw an error).}

\item{nas}{A flag indicating whether missing values are permitted.}

\item{floored}{A flag indicating whether x should be floored.}

\item{sorted}{A flag indicating whether the vector must be sorted.}

\item{unique}{A flag indicating whether the values must be unique.}

\item{complete}{A flag indicating whether x should be complete.}

\item{units}{A string of the units.}

\item{tz}{A string of the time zone.}

\item{exclusive}{A flag indicating whether other columns are not permitted.}

\item{order}{A flag indicating whether the columns have to occur in the same order as values.}

\item{x_name}{A string of the name of the object x.}

\item{error}{A flag indicating whether to throw an informative error or
immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
A dts data frame is a data frame with a column of Date or POSIXct values.
}
\examples{
check_dts(dts_data)
}
