% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts-package.R
\docType{package}
\name{dts-package}
\alias{dts}
\alias{dts-package}
\title{dts: Manipulate Date/POSIXct Time Series Data Frames}
\description{
Manipulates data frames with a Date or POSIXct column and one or more
    double columns representing time series values.
}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)

}
\keyword{internal}
