% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{dts_aggregate}
\alias{dts_aggregate}
\title{Aggregate}
\usage{
dts_aggregate(x, date_time = "DateTime", value = "Value",
  units = dttr::dtt_units(x[[date_time]]), .fun = mean, ...)
}
\arguments{
\item{.fun}{A function which returns a scalar.}

\item{...}{Optional Additional arguments passed to .fun.}
}
\value{
A data frame
}
\description{
Aggregate
}
\examples{
dts_aggregate(dts_data, units = "years", na.rm = TRUE)
}
