% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{dts_aggregate}
\alias{dts_aggregate}
\title{Aggregate}
\usage{
dts_aggregate(
  x,
  dtt = "DateTime",
  colname = dts_colnames(x),
  units = dttr::dtt_units(x[[dtt]]),
  .dts_fun = mean,
  ...
)
}
\arguments{
\item{x}{A dts data frame}

\item{dtt}{A string specifying the column with the Date or POSIXct values.}

\item{colname}{A character vector specifying the column(s) with the other values.}

\item{units}{A string of the units.}

\item{.dts_fun}{A function which returns a scalar.}

\item{...}{Optional Additional arguments passed to .dts_fun.}
}
\value{
A data frame
}
\description{
Aggregate
}
\examples{
dts_aggregate(dts_data, units = "years", na.rm = TRUE)
}
