% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound.R
\name{dts_bound}
\alias{dts_bound}
\title{Bound}
\usage{
dts_bound(x, dtt = "DateTime", colname = "Value", bounds = c(-Inf,
  Inf), adjust = FALSE)
}
\arguments{
\item{x}{A dts data frame}

\item{dtt}{A string specifying the column with the Date or POSIXct values.}

\item{colname}{A character vector specifying the column(s) with the other values.}

\item{bounds}{A sorted double vector of the lower and upper permitted bounds.}

\item{adjust}{A flag indicating whether to adjust values outside the bounds to the permitted bounds
or whether to set to NA (the default).}
}
\value{
A data frame
}
\description{
Values outside the bounds are set to missing values.
}
\examples{
dts_bound(dts_data[1:5,], bounds = c(0,Inf))
}
