% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{dts_complete}
\alias{dts_complete}
\title{Completes a data time series}
\usage{
dts_complete(
  x,
  dtt = "DateTime",
  colname = dts_colnames(x),
  from = min(x[[dtt]]),
  to = max(x[[dtt]]),
  floor = TRUE,
  unique = TRUE,
  sort = TRUE,
  units = dttr2::dtt_units(x[[dtt]]),
  .dts_fun = mean,
  ...
)
}
\arguments{
\item{x}{A date/time vector.}

\item{dtt}{A string specifying the column with the Date or POSIXct values.}

\item{colname}{A character vector specifying the column(s) with the other
values.}

\item{from}{A date/time scalar of the start.}

\item{to}{A date/time scalar of the end.}

\item{floor}{A flag indicating whether to floor the values.}

\item{unique}{A flag specifying whether to only return unique values.}

\item{sort}{A flag specifying whether to sort the vector.}

\item{units}{A string of the time units. The possible values are "secs",
"minutes", "hours", "days" or "weeks".}

\item{.dts_fun}{A function which returns a scalar.}

\item{...}{Unused.}
}
\value{
The completed data time series.
}
\description{
Completes a data time series
}
\examples{
dts_complete(dts_data[c(1,3),])
}
