% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dts_complete}
\alias{dts_complete}
\title{Complete}
\usage{
dts_complete(x)
}
\arguments{
\item{x}{A POSIXct object.}
}
\value{
A flag indicating whether complete
}
\description{
Tests whether a POSIXct object is complete.
}
\examples{
dts_complete(dts_data$DateTime[c(1,2,3)])
dts_complete(dts_data$DateTime[c(1,3)])
dts_complete(dts_data$DateTime[c(1,3,2)]) 
}
