% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{dts_complete}
\alias{dts_complete}
\title{Bound}
\usage{
dts_complete(x, date_time = "DateTime", value = "Value",
  units = dttr::dtt_units(x[[date_time]]))
}
\arguments{
\item{x}{A dts data frame}

\item{date_time}{A string specifying the column with the Date or POSIXct values.}

\item{value}{A string specifying the column with the double values.}

\item{units}{A string of the time units.}
}
\value{
A data frame
}
\description{
Values outside the bounds are set to missing values.
}
\examples{
dts_complete(dts_data[c(1,3),])
}
