% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completed.R
\name{dts_completed}
\alias{dts_completed}
\title{Test if the DateTime in a DateTime-Series is complete.}
\usage{
dts_completed(
  x,
  dtt = "DateTime",
  floored = TRUE,
  unique = TRUE,
  sorted = TRUE,
  units = dttr::dtt_units(x[[dtt]])
)
}
\arguments{
\item{x}{A dts data frame}

\item{dtt}{A string specifying the column with the Date or POSIXct values.}

\item{floored}{A flag indicating whether x should be floored.}

\item{unique}{A flag indicating whether the values must be unique.}

\item{sorted}{A flag indicating whether the vector must be sorted.}

\item{units}{A string of the units.}
}
\value{
A flag indicating whether complete.
}
\description{
Test if the DateTime in a DateTime-Series is complete.
}
\examples{
dts_completed(dts_data[c(1,3),])
}
\seealso{
\code{\link[dttr]{dtt_completed}}
}
