% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay.R
\name{dts_delay}
\alias{dts_delay}
\title{Delay a data time series}
\usage{
dts_delay(x, dtt = "DateTime", colname = dts_colnames(x),
  units = dttr::dtt_units(x[[dtt]]), delay = 0L)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{dtt}{A string specifying the column with the Date or POSIXct values.}

\item{colname}{A character vector specifying the column(s) with the other values.}

\item{units}{A string of the units.}

\item{delay}{An int (integer scalar) of the delay.}
}
\value{
The completed data time series.
}
\description{
Delay a data time series
}
\examples{
dts_delay(dts_data[1:4,], delay = 2)
}
