% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{dts_interpolate}
\alias{dts_interpolate}
\title{Interpolate}
\usage{
dts_interpolate(x, date_time = "DateTime", value = "Value",
  max_span = 10L, method = "linear", step = 0.5)
}
\arguments{
\item{x}{A dts data frame}

\item{date_time}{A string specifying the column with the Date or POSIXct values.}

\item{value}{A string specifying the column with the double values.}

\item{max_span}{An integer of the maximum span to interpolate.}

\item{method}{A string of "linear" or "constant" indicating the method to use.}

\item{step}{A proportion specifing the compromise between 
left- and right- continuous step function.}
}
\value{
A data frame
}
\description{
Fill in missing values by linear interpolation.
}
\examples{
dts_interpolate(dts_data[1:5,])
}
