% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{dts_regress}
\alias{dts_regress}
\title{Regress}
\usage{
dts_regress(
  x,
  dtt = "DateTime",
  colname = c("Value", "Value2"),
  intercept = TRUE,
  min_gap = 0L,
  min_n = 4L
)
}
\arguments{
\item{x}{A dts data frame}

\item{dtt}{A string specifying the column with the Date or POSIXct values.}

\item{colname}{A character vector specifying the column(s) with the other values.}

\item{intercept}{A flag indicating whether an intercept should be estimated (or fixed at 0)}

\item{min_gap}{An integer of the minimum gap to leave between regressed and existing values.}

\item{min_n}{A count of the minimum of data points to require.}
}
\value{
A data frame
}
\description{
Fill in missing values by linear regression.
}
\examples{
dts_regress(dts_data)[1:5,]
}
