% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_water_bath.R
\name{find_water_bath}
\alias{find_water_bath}
\title{find_water_bath}
\usage{
find_water_bath(x, type = "correlation", ...)
}
\arguments{
\item{x}{the dts_long dataset}

\item{type}{use correlation or absolute values of temperature}

\item{...}{arguments to pass to correlate_with_temperature}
}
\description{
find_water_bath
}
