% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_convolve.R
\name{fit_convolve}
\alias{fit_convolve}
\title{fit_convolve}
\usage{
fit_convolve(x, cool_mult = 0.5, time_var = "start", n_knots = NULL)
}
\arguments{
\item{x}{}

\item{cool_mult}{}

\item{time_var}{}

\item{n_knots}{}
}
\value{

}
\description{
fit_convolve
}
