% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_convolve.R
\name{fit_convolve}
\alias{fit_convolve}
\title{fit_convolve}
\usage{
fit_convolve(x, n_knots = 10, cool_mult = 0.5)
}
\arguments{
\item{x}{}

\item{n_knots}{}

\item{cool_mult}{}
}
\value{

}
\description{
fit_convolve
}
