% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\alias{generate_report.character}
\alias{generate_report.dts_long}
\title{generate_report}
\usage{
generate_report(files_dir, output_dir, ...)

\method{generate_report}{character}(files_dir, output_dir)

\method{generate_report}{dts_long}(files_dir, output_dir)
}
\arguments{
\item{files_dir}{}

\item{output_dir}{}
}
\value{

}
\description{
generate_report
}
