% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\alias{generate_report.character}
\alias{generate_report.dts_long}
\title{generate_report}
\usage{
generate_report(input, output_dir, fmt = "bookdown::gitbook", ...)

\method{generate_report}{character}(input, output_dir, fmt = "bookdown::gitbook", n_cores = 1, ...)

\method{generate_report}{dts_long}(input, output_dir, fmt = "bookdown::gitbook", dir_name, ...)
}
\arguments{
\item{input}{folder or dts data already parsed}

\item{output_dir}{}

\item{fmt}{}

\item{...}{heating_power, heating_type}
}
\value{

}
\description{
generate_report
}
