% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temperature_breakpoint.R
\name{get_temperature_breakpoint}
\alias{get_temperature_breakpoint}
\alias{get_temperature_breakpoint.dts_long}
\title{get_temperature_breakpoint}
\usage{
get_temperature_breakpoint(
  x,
  datetime_val,
  n_values = 1000,
  n_buffer = 2,
  temp_col = "temperature"
)

\method{get_temperature_breakpoint}{dts_long}(
  x,
  datetime_val,
  n_values = 1000,
  n_buffer = 2,
  temp_col = "temperature"
)
}
\arguments{
\item{x}{}

\item{datetime_val}{}

\item{n_values}{}

\item{n_buffer}{}

\item{temp_col}{}
}
\value{

}
\description{
get_temperature_breakpoint
}
