% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heating_distance.R
\name{heating_distance}
\alias{heating_distance}
\title{heating_distance}
\usage{
heating_distance(x, n = 50, heating_type = "both")
}
\arguments{
\item{x}{}

\item{n}{}

\item{heating_type}{}
}
\value{

}
\description{
heating_distance
}
