% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heating_time.R
\name{heating_time}
\alias{heating_time}
\title{heating_time}
\usage{
heating_time(x, n = -10, heating_type = "both", distance_range = NULL)
}
\arguments{
\item{x}{}

\item{n}{}

\item{heating_type}{}
}
\value{

}
\description{
heating_time
}
