% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_distances}
\alias{plot_distances}
\title{plot_distances}
\usage{
plot_distances(dts, n_traces = 10)
}
\arguments{
\item{dts}{the dts set}

\item{n_traces}{number of traces to plot}

\item{trim_min}{minimum temperature values to set to NA}
}
\value{
a plotly heatmap
}
\description{
plot_distances
}
