% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_distances}
\alias{plot_distances}
\title{plot_distances}
\usage{
plot_distances(dts, n = 10)
}
\arguments{
\item{dts}{the dts set}

\item{n}{number of traces to plot}
}
\value{
a plotly heatmap
}
\description{
plot_distances
}
