% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_times}
\alias{plot_times}
\title{plot_times}
\usage{
plot_times(dts, n_traces = 10)
}
\arguments{
\item{dts}{the dts set}

\item{n_times}{number of traces to plot}
}
\value{
a plotly heatmap
}
\description{
plot_times
}
