% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dts_xml.R
\name{read_dts_folder}
\alias{read_dts_folder}
\title{read_dts_folder}
\usage{
read_dts_folder(directory, out_path, ...)
}
\arguments{
\item{directory}{location of the xml files}

\item{...}{arguments to pass to mclapply}
}
\description{
read_dts_folder
}
