% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dts_xml.R
\name{read_dts_xml}
\alias{read_dts_xml}
\title{read_dts_xml}
\usage{
read_dts_xml(file_path, trim = TRUE)
}
\arguments{
\item{file_path}{path to the file}

\item{trim}{whether to trim the length of the cable}
}
\value{
dts table of results
}
\description{
read_dts_xml
}
