% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dts_xml.R
\name{read_dts_xml}
\alias{read_dts_xml}
\title{read_dts_xml}
\usage{
read_dts_xml(in_dir, n_cores, date_format = "R", max_files = Inf)
}
\arguments{
\item{in_dir}{path to the input folder}

\item{n_cores}{number of cores to use for parallel processing}

\item{date_format}{'matlab' or 'R' format}

\item{max_files}{maximum number of files to read}
}
\value{
dts list of results
}
\description{
read_dts_xml
}
