% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_time}
\alias{select_time}
\alias{select_time.dts_long}
\alias{select_time.data.table}
\alias{select_time_ind}
\title{select_time}
\usage{
select_time(x, times, ...)

\method{select_time}{dts_long}(x, times)

\method{select_time}{data.table}(x, times)

select_time_ind(x, inds, ...)
}
\arguments{
\item{x}{data read from a dts}

\item{...}{}
}
\description{
select_time

select_time_ind
}
