% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_time}
\alias{subset_time}
\alias{subset_time.dts_long}
\alias{subset_time.data.table}
\title{subset_time}
\usage{
subset_time(x, begin, end, ...)

\method{subset_time}{dts_long}(x, begin, end)

\method{subset_time}{data.table}(x, begin, end)
}
\arguments{
\item{x}{data read from a dts}

\item{...}{}
}
\value{

}
\description{
subset_time
}
