% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermal_conductivity.R
\name{thermal_conductivity}
\alias{thermal_conductivity}
\title{thermal_conductivity}
\usage{
thermal_conductivity(x, power)
}
\arguments{
\item{x}{}

\item{power}{}
}
\value{

}
\description{
thermal_conductivity
}
