% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_breakpoints.R
\name{time_breakpoints}
\alias{time_breakpoints}
\alias{time_breakpoints.default}
\alias{time_breakpoints.matrix}
\title{time_breakpoints}
\usage{
time_breakpoints(x, shift, col_name)

\method{time_breakpoints}{default}(x, shift = -10, col_name = "temperature")

\method{time_breakpoints}{matrix}(x, shift = -10, col_name = "temperature")
}
\arguments{
\item{x}{data read from a dts}

\item{shift}{distance between differences in number of samples}

\item{col_name}{}
}
\value{

}
\description{
time_breakpoints
}
