% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dts_xml_3.R
\name{to_duckdb_2}
\alias{to_duckdb_2}
\title{to_duckdb_2}
\usage{
to_duckdb_2(
  file_path,
  db_dir,
  n_cores = 1L,
  max_files = Inf,
  return_stokes = FALSE,
  trim = TRUE
)
}
\arguments{
\item{trim}{}
}
\description{
to_duckdb_2
}
