% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_duckdb.R, R/to_matrix.R
\name{to_duckdb}
\alias{to_duckdb}
\alias{to_matrix}
\alias{to_matrix.dts}
\alias{to_matrix.data.table}
\alias{to_matrix.dts_long}
\title{to_duckdb}
\usage{
to_duckdb(x, dbdir, n_cores = 1)

to_matrix(x, ...)

\method{to_matrix}{dts}(x, ...)

\method{to_matrix}{data.table}(x, col_name = "temperature")

\method{to_matrix}{dts_long}(x, col_name = "temperature")
}
\arguments{
\item{x}{data read from a dts}

\item{...}{}
}
\value{
saves a duckdb database
}
\description{
to_duckdb

to_matrix
}
