% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_matrix.R
\name{to_matrix}
\alias{to_matrix}
\alias{to_matrix.dts}
\alias{to_matrix.data.table}
\alias{to_matrix.dts_long}
\title{to_matrix}
\usage{
to_matrix(x, ...)

\method{to_matrix}{dts}(x, ...)

\method{to_matrix}{data.table}(x, col_name = "temperature")

\method{to_matrix}{dts_long}(x, col_name = "temperature")
}
\arguments{
\item{x}{data read from a dts}

\item{...}{}
}
\value{

}
\description{
to_matrix
}
