% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dts_xml.R
\name{write_dts}
\alias{write_dts}
\title{write_dts}
\usage{
write_dts(in_path, out_path, n_cores = 1)
}
\arguments{
\item{in_path}{input directory}

\item{out_path}{output directory}

\item{n_cores}{how many cores}
}
\description{
write_dts
}
